+++ BALSAMIC/models/config.py
@@ -1,16 +1,10 @@
-import glob
-import logging
+"""Balsamic analysis config case models."""
 import re
+from glob import glob
 from pathlib import Path
 from typing import Annotated, Dict, List, Optional
 
-from pydantic import (
-    AfterValidator,
-    BaseModel,
-    ConfigDict,
-    field_validator,
-    model_validator,
-)
+from pydantic import AfterValidator, BaseModel, field_validator, model_validator
 
 from BALSAMIC import __version__ as balsamic_version
 from BALSAMIC.constants.analysis import (
@@ -25,104 +19,49 @@ from BALSAMIC.constants.analysis import (
     SequencingType,
     WorkflowSolution,
 )
+from BALSAMIC.models.params import QCModel
+from BALSAMIC.models.validators import is_dir, is_file
 
-LOG = logging.getLogger(__name__)
+
+class FastqInfoModel(BaseModel):
+    """Holds filepaths for forward and reverse reads for a fastq_pattern."""
+
+    fwd: Annotated[str, AfterValidator(is_file)]
+    rev: Annotated[str, AfterValidator(is_file)]
+    fwd_resolved: Annotated[Optional[str], AfterValidator(is_file)] = None
+    rev_resolved: Annotated[Optional[str], AfterValidator(is_file)] = None
 
 
-class VCFAttributes(BaseModel):
-    """General purpose filter to manage various VCF attributes
-
-    This class handles three parameters for the purpose filtering variants
-    based on a tag_values, filter_name, and which field in VCF.
-
-    E.g. AD=VCFAttributes(tag_value=5, filter_name="balsamic_low_tumor_ad", field="INFO")
-    A value of 5 from INFO field and filter_name will be balsamic_low_tumor_ad
+class SampleInstanceModel(BaseModel):
+    """Holds attributes for samples used in analysis.
 
     Attributes:
-        tag_value: float
-        filter_name: str
-        field: str
+        type: Field(str): sample type [tumor, normal]
+        name: Field(str): sample name
+        fastq_info: Field(dict): fastq patterns: paths to forward and reverse fastqs
     """
 
-    tag_value: float
-    filter_name: str
-    field: str
+    type: SampleType
+    name: str
+    fastq_info: Dict[str, FastqInfoModel]
 
 
-class VarCallerFilter(BaseModel):
-    """General purpose for variant caller filters
-
-    This class handles attributes and filter for variant callers
-
+class PanelModel(BaseModel):
+    """Holds attributes of PANEL BED file if provided
     Attributes:
-        AD: VCFAttributes (required); minimum allelic depth
-        AF_min: VCFAttributes (optional); minimum allelic fraction
-        AF_max: VCFAttributes (optional); maximum allelic fraction
-        MQ: VCFAttributes (optional); minimum mapping quality
-        DP: VCFAttributes (optional); minimum read depth
-        pop_freq: VCFAttributes (optional); maximum gnomad allele frequency
-        pop_freq_umi: VCFAttributes (optional); maximum gnomad_af for UMI workflow
-        strand_reads: VCFAttributes (optional); minimum strand specific read counts
-        qss: VCFAttributes (optional); minimum sum of base quality scores
-        sor: VCFAttributes (optional); minimum symmetrical log-odds ratio
-        swegen_snv_freq: VCFAttributes (optional); maximum swegen snv allele frequency
-        swegen_sv_freq: VCFAttributes (optional); maximum swegen sv allele frequency
-        loqusdb_clinical_snv_freq: VCFAttributes (optional); maximum loqusdb clinical snv allele frequency
-        loqusdb_clinical_sv_freq: VCFAttributes (optional); maximum loqusdb clinical sv allele frequency
-        varcaller_name: str (required); variant caller name
-        filter_type: str (required); filter name for variant caller
-        analysis_type: str (required); analysis type e.g. tumor_normal or tumor_only
-        description: str (required); comment section for description
-    """
-
-    AD: Optional[VCFAttributes] = None
-    AF_min: Optional[VCFAttributes] = None
-    AF_max: Optional[VCFAttributes] = None
-    MQ: Optional[VCFAttributes] = None
-    DP: Optional[VCFAttributes] = None
-    pop_freq: Optional[VCFAttributes] = None
-    pop_freq_umi: Optional[VCFAttributes] = None
-    strand_reads: Optional[VCFAttributes] = None
-    qss: Optional[VCFAttributes] = None
-    sor: Optional[VCFAttributes] = None
-    swegen_snv_freq: Optional[VCFAttributes] = None
-    swegen_sv_freq: Optional[VCFAttributes] = None
-    loqusdb_clinical_snv_freq: Optional[VCFAttributes] = None
-    loqusdb_clinical_sv_freq: Optional[VCFAttributes] = None
-    varcaller_name: str
-    filter_type: str
-    analysis_type: str
-    description: str
-
-
-class QCModel(BaseModel):
-    """Contains settings for quality control and pre-processing
-    Attributes:
-        picard_rmdup : Field(bool); whether duplicate removal is to be applied in the workflow
-        adapter : Field(str(AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT)); adapter sequence to trim
-        quality_trim : Field(bool); whether quality trimming it to be performed in the workflow
-        adapter_trim : Field(bool); whether adapter trimming is to be performed in the workflow
-        umi_trim : Field(bool); whether UMI trimming is to be performed in the workflow
-        min_seq_length : Field(str(int)); minimum sequence length cutoff for reads
-        umi_trim_length : Field(str(int)); length of UMI to be trimmed from reads
-        n_base_limit : Field(str(int)); supports filtering by limiting the N base number
+        capture_kit : Field(str(Path)); string representation of path to PANEL BED file
+        chrom : Field(list(str)); list of chromosomes in PANEL BED
+        pon_cnn: Field(optional); Path where PON reference .cnn file is stored
 
     Raises:
         ValueError:
-            When the input in min_seq_length and umi_trim_length cannot
-            be interpreted as integer and coerced to string
+            When capture_kit argument is set, but is not a valid path
 
     """
 
-    model_config = ConfigDict(coerce_numbers_to_str=True)
-    picard_rmdup: bool = False
-    adapter: str = "AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT"
-    quality_trim: bool = True
-    adapter_trim: bool = False
-    umi_trim: bool = False
-    min_seq_length: str = "25"
-    umi_trim_length: str = "5"
-    n_base_limit: str = "50"
+    capture_kit: Annotated[Optional[str], AfterValidator(is_file)] = None
+    chrom: Optional[List[str]] = None
+    pon_cnn: Annotated[Optional[str], AfterValidator(is_file)] = None
 
 
 class VarcallerAttribute(BaseModel):
@@ -164,24 +103,6 @@ class VCFModel(BaseModel):
     tiddit: VarcallerAttribute
     cnvpytor: VarcallerAttribute
     svdb: VarcallerAttribute
-
-
-def is_file(file_path: Optional[str]) -> str:
-    """Validate file path existence."""
-    if file_path:
-        if Path(file_path).is_file():
-            return file_path
-        raise ValueError(f"The supplied file path {file_path} does not exist")
-    return file_path
-
-
-def is_dir(dir_path: Optional[str]) -> str:
-    """Validate directory path existence."""
-    if dir_path:
-        if Path(dir_path).is_dir():
-            return dir_path
-        raise ValueError(f"The supplied directory path {dir_path} does not exist")
-    return dir_path
 
 
 class AnalysisModel(BaseModel):
@@ -245,47 +166,6 @@ class AnalysisModel(BaseModel):
                 f"The provided PON version ({pon_version}) does not follow the defined syntax (v<int>)"
             )
         return pon_version
-
-
-class FastqInfoModel(BaseModel):
-    """Holds filepaths for forward and reverse reads for a fastq_pattern."""
-
-    fwd: Annotated[str, AfterValidator(is_file)]
-    rev: Annotated[str, AfterValidator(is_file)]
-    fwd_resolved: Annotated[Optional[str], AfterValidator(is_file)] = None
-    rev_resolved: Annotated[Optional[str], AfterValidator(is_file)] = None
-
-
-class SampleInstanceModel(BaseModel):
-    """Holds attributes for samples used in analysis.
-
-    Attributes:
-        type: Field(str): sample type [tumor, normal]
-        name: Field(str): sample name
-        fastq_info: Field(dict): fastq patterns: paths to forward and reverse fastqs
-    """
-
-    type: SampleType
-    name: str
-    fastq_info: Dict[str, FastqInfoModel]
-
-
-class PanelModel(BaseModel):
-    """Holds attributes of PANEL BED file if provided
-    Attributes:
-        capture_kit : Field(str(Path)); string representation of path to PANEL BED file
-        chrom : Field(list(str)); list of chromosomes in PANEL BED
-        pon_cnn: Field(optional); Path where PON reference .cnn file is stored
-
-    Raises:
-        ValueError:
-            When capture_kit argument is set, but is not a valid path
-
-    """
-
-    capture_kit: Annotated[Optional[str], AfterValidator(is_file)] = None
-    chrom: Optional[List[str]] = None
-    pon_cnn: Annotated[Optional[str], AfterValidator(is_file)] = None
 
 
 class ConfigModel(BaseModel):
@@ -394,7 +274,7 @@ class ConfigModel(BaseModel):
         fastq_path = values["analysis"]["fastq_path"]
 
         # Get a set of all fastq files in fastq-directory
-        fastqs_in_fastq_path = set(glob.glob(f"{fastq_path}/*fastq.gz"))
+        fastqs_in_fastq_path = set(glob(f"{fastq_path}/*fastq.gz"))
 
         # Look for fastqs in sample dict
         fastqs_assigned = set(get_all_fwd_rev_values(values["samples"]))
@@ -529,135 +409,3 @@ class ConfigModel(BaseModel):
             final_bam_suffix = "dedup.realign"
 
         return f"{bam_dir}{sample_type}.{sample_name}.{final_bam_suffix}.bam"
-
-
-class UMIParamsCommon(BaseModel):
-    """This class defines the common params settings used as constants across various rules in UMI workflow.
-
-    Attributes:
-        align_format: str (required); output alignment format. eg. 'BAM'
-        align_header: str (required); header line appended to the aligned BAM output
-        align_intbases: int; input bases in each batch regardless of threads, for reproducibility
-        filter_tumor_af: float (required); settings to filter minimum allelic frequency
-    """
-
-    align_header: str
-    align_intbases: int
-    filter_tumor_af: float
-
-
-class UMIParamsUMIextract(BaseModel):
-    """This class defines the params settings used as constants in UMI workflow-rule umextract.
-
-    Attributes:
-        read_structure: str (required); settings to define UMI read structure
-    """
-
-    read_structure: str = "-d, 'rs1,rs2'"
-
-
-class UMIParamsConsensuscall(BaseModel):
-    """This class defines the params settings used as constants in UMI workflow-rule consensuscall.
-
-    Attributes:
-        align_format: str (required); output alignment format. eg. 'BAM'
-            filter_minreads: str (required); settings to filter consensus tags based on group size
-        tag: str; Logic UMI tag
-    """
-
-    align_format: str = "BAM"
-    filter_minreads: str = "3,1,1"
-    tag: str = "XR"
-
-
-class UMIParamsTNscope(BaseModel):
-    """This class defines the params settings used as constants in UMI workflow- rule tnscope.
-
-    Attributes:
-        algo: str; choice of sentieon varcall algorithm. eg. 'TNscope'
-        disable_detect: str; disable variant detector. eg 'sv' or 'snv_indel'
-        filter_tumor_af: float (required); minimum allelic frequency to detect
-        min_tumorLOD: int (required); minimum tumor log odds in the final call of variants
-        init_tumorLOD: float (required); minimum tumor log odds in the initial pass calling variants
-        error_rate: int (required); allow error-rate to consider in calling
-        prunefactor: int (required); pruning factor in the kmer graph
-        padding: int(required); amount to pad bed interval regions
-    """
-
-    algo: str
-    init_tumorLOD: float
-    min_tumorLOD: int
-    error_rate: int
-    prunefactor: int
-    padding: int
-    disable_detect: str
-
-
-class ParamsVardict(BaseModel):
-    """This class defines the params settings used as constants in vardict rule.
-
-    Attributes:
-        allelic_frequency: float (required); minimum allelic frequency to detect
-        max_pval: float (required); the maximum p-value. Vardict default: 0.05
-        max_mm: float (required); the maximum mean mismatches allowed. Vardict default: 5.25
-        column_info: str (required); set of vardict filters for passing final variants
-    """
-
-    allelic_frequency: float
-    max_pval: float
-    max_mm: float
-    column_info: str
-
-
-class ParamsCommon(BaseModel):
-    """This class defines the common params settings used as constants across various rules in balsamic workflow.
-
-    Attributes:
-        pcr_model: str (required). PCR indel model used to weed out false positive indels. Eg: none- PCR free samples.
-        align_header: str (required); header line appended to the aligned BAM output
-        min_mapq: int (required); minimum mapping quality score. Eg: 20- probability of mapping random read at 99% accuracy
-        picard_fixmate: str (required), fix read mate information in bam file
-        picard_RG_normal: str (required); replace readgroups in normal bam file
-        picard_RG_tumor: str (required); replace readgroups in tumor bam file
-    """
-
-    align_header: str
-    pcr_model: str
-    min_mapq: int
-    picard_fixmate: str
-    picard_RG_normal: str
-    picard_RG_tumor: str
-
-
-class ParamsVEP(BaseModel):
-    """This class defines the params settings used as constants in vep rule.
-
-    Attributes:
-        vep_filters: str (required); set of choosen options for processing vep annotated vcf file
-    """
-
-    vep_filters: str
-
-
-class BalsamicWorkflowConfig(BaseModel):
-    """Defines set of rules in balsamic workflow
-
-    Handles attributes for corresponding rules.
-
-    Attributes:
-        common: global params defined across all rules in balsamic workflow
-        umicommon: global params defined across specific rules in UMI workflow
-        vep: global params defined in the rule vep
-        vardict: params defined in the rule vardict
-        umiextract : params defined in the rule sentieon_umiextract
-        umiconsensuscall: params defined in the rule sentieon_consensuscall
-        tnscope_umi: params defined in the rule sentieon_tnscope_umi
-    """
-
-    common: ParamsCommon
-    vardict: ParamsVardict
-    vep: ParamsVEP
-    umicommon: UMIParamsCommon
-    umiextract: UMIParamsUMIextract
-    umiconsensuscall: UMIParamsConsensuscall
-    tnscope_umi: UMIParamsTNscope
